Move batch-level selector dispatch into FilterSpec.matchBatch
FilterProcessor no longer casts EventGroupIdentifier manually. Each
FilterSpec variant implements matchBatch returning BatchMatchResult
(Skip | Matched), with batchEntityKeysOverlap moving into ByEntityKeys
as a private helper. MissingBatchEntityKeysException also moves to
FilterSpec.kt since it is thrown by matchBatch.
