+++ src/main/kotlin/org/wfanet/measurement/edpaggregator/resultsfulfiller/FilterProcessor.kt
@@ -25,28 +25,13 @@ import org.wfanet.measurement.edpaggregator.v1alpha.LabeledImpression
 import org.wfanet.measurement.eventdataprovider.eventfiltration.EventFilters
 
 /**
- * Thrown by [FilterProcessor.processBatch] when the active selector is [FilterSpec.ByEntityKeys]
- * but the incoming [EventBatch] carries no `entity_keys`.
- *
- * Indicates a wiring/data invariant violation upstream (the blob's `BlobDetails.entity_keys` was
- * not populated despite the consumer requesting entity-key filtering).
- */
-class MissingBatchEntityKeysException :
-  IllegalStateException(
-    "ByEntityKeys filter requires the batch to carry entity_keys, but batch.entityKeys is empty"
-  )
-
-/**
  * Filter processor for event filtering with CEL expressions and time ranges.
  *
  * This processor compiles a CEL expression once and reuses it for batch processing. It processes
  * events sequentially through the CEL filter and identifies matching events.
  *
- * The processor dispatches on [FilterSpec]'s variant to choose the EventGroup selector:
- * - [FilterSpec.ByEventGroupReferenceIds] uses the legacy `eventGroupReferenceId` batch-level
- *   match.
- * - [FilterSpec.ByEntityKeys] uses a batch-level entity-key intersection short-circuit followed by
- *   a per-event entity-key check.
+ * Batch-level selector dispatch (reference-id vs entity-key) is handled by [FilterSpec.matchBatch],
+ * which returns a [BatchMatchResult].
  *
  * @param filterSpec immutable specification containing all filtering criteria
  */
@@ -63,24 +48,12 @@ class FilterProcessor<T : Message>(
       EventFilters.compileProgram(eventDescriptor, filterSpec.celExpression)
     }
 
-  /**
-   * Start time instant from the collection interval.
-   *
-   * Pre-computed to avoid repeated conversion from protobuf Timestamp to Java Instant during batch
-   * processing.
-   */
   private val startInstant: Instant =
     Instant.ofEpochSecond(
       filterSpec.collectionInterval.startTime.seconds,
       filterSpec.collectionInterval.startTime.nanos.toLong(),
     )
 
-  /**
-   * End time instant from the collection interval.
-   *
-   * Pre-computed to avoid repeated conversion from protobuf Timestamp to Java Instant during batch
-   * processing.
-   */
   private val endInstant: Instant =
     Instant.ofEpochSecond(
       filterSpec.collectionInterval.endTime.seconds,
@@ -91,11 +64,7 @@ class FilterProcessor<T : Message>(
    * Processes a batch of events and returns the matching events.
    *
    * Applies filtering in the following order for optimal performance:
-   * 1. Selector check, dispatched on [FilterSpec] variant:
-   *     - [FilterSpec.ByEventGroupReferenceIds]: batch-level `event_group_reference_id` match.
-   *     - [FilterSpec.ByEntityKeys]: batch-level entity-key intersection short-circuit. Throws
-   *       [MissingBatchEntityKeysException] when the batch carries no entity_keys (data invariant
-   *       violation).
+   * 1. Batch-level selector check via [FilterSpec.matchBatch].
    * 2. Batch time range overlap check (fast, avoids processing entire batch).
    * 3. Per-event time range filter.
    * 4. Per-event entity-key intersection (only for [FilterSpec.ByEntityKeys]).
@@ -105,7 +74,7 @@ class FilterProcessor<T : Message>(
    * @return An EventBatch containing events that match all configured filters. The batch may be
    *   empty if no events match the criteria.
    * @throws MissingBatchEntityKeysException when [filterSpec] is [FilterSpec.ByEntityKeys] and
-   *   [batch.entityKeys] is empty.
+   *   [batch.eventGroupIdentifier] is not [EventGroupIdentifier.ByEntityKeys].
    */
   fun processBatch(batch: EventBatch<T>): EventBatch<T> {
     if (batch.events.isEmpty()) {
@@ -113,29 +82,11 @@ class FilterProcessor<T : Message>(
     }
 
     val entityKeyFilter: Set<LabeledImpression.EntityKey>? =
-      when (val spec = filterSpec) {
-        is FilterSpec.ByEventGroupReferenceIds -> {
-          val identifier =
-            batch.eventGroupIdentifier as? EventGroupIdentifier.ByReferenceId
-              ?: return emptyBatchLike(batch)
-          if (!spec.eventGroupReferenceIds.contains(identifier.refId)) {
-            return emptyBatchLike(batch)
-          }
-          null
-        }
-        is FilterSpec.ByEntityKeys -> {
-          val identifier =
-            batch.eventGroupIdentifier as? EventGroupIdentifier.ByEntityKeys
-              ?: throw MissingBatchEntityKeysException()
-          if (identifier.entityKeys.isEmpty()) throw MissingBatchEntityKeysException()
-          if (!batchEntityKeysOverlap(identifier.entityKeys, spec.entityKeys)) {
-            return emptyBatchLike(batch)
-          }
-          spec.entityKeys
-        }
+      when (val result = filterSpec.matchBatch(batch.eventGroupIdentifier)) {
+        is BatchMatchResult.Skip -> return emptyBatchLike(batch)
+        is BatchMatchResult.Matched -> result.entityKeyFilter
       }
 
-    // Fast batch-level time range check: skip entire batch if no overlap
     if (!batchTimeRangeOverlaps(batch)) {
       return emptyBatchLike(batch)
     }
@@ -161,12 +112,6 @@ class FilterProcessor<T : Message>(
     )
   }
 
-  /**
-   * Returns an empty `EventBatch` carrying the same metadata as [batch].
-   *
-   * Used to short-circuit a batch without dropping its identifying metadata so downstream
-   * accounting can still distinguish per-batch outputs.
-   */
   private fun emptyBatchLike(batch: EventBatch<T>): EventBatch<T> {
     return EventBatch(
       emptyList(),
@@ -176,60 +121,19 @@ class FilterProcessor<T : Message>(
     )
   }
 
-  /**
-   * Checks if the batch's time range overlaps with the filter's collection interval.
-   *
-   * This is a fast batch-level check to avoid processing events when the entire batch is outside
-   * the collection interval.
-   */
   private fun batchTimeRangeOverlaps(batch: EventBatch<T>): Boolean {
     return !batch.maxTime.isBefore(startInstant) && batch.minTime.isBefore(endInstant)
   }
 
-  /**
-   * Checks if an event's timestamp falls within the collection interval.
-   *
-   * Uses a half-open interval [start, end) where the start time is inclusive and the end time is
-   * exclusive.
-   */
   private fun isEventInTimeRange(event: LabeledEvent<T>): Boolean {
     val eventTime = event.timestamp
     return !eventTime.isBefore(startInstant) && eventTime.isBefore(endInstant)
   }
 
-  /**
-   * Checks whether [event]'s `entityKeys` intersect [filter].
-   *
-   * An event with empty `entityKeys` always fails this check (cannot match a non-empty filter).
-   */
   private fun eventMatchesEntityKeyFilter(
     event: LabeledEvent<T>,
     filter: Set<LabeledImpression.EntityKey>,
   ): Boolean {
     return event.entityKeys.any { it in filter }
   }
-
-  /**
-   * Checks whether any [org.wfanet.measurement.edpaggregator.v1alpha.EntityKeyGroup] on [batch]
-   * contains an `(entity_type, entity_id)` pair that matches [filter].
-   *
-   * Pre-flattens the batch's grouped entity keys into a `Set<EntityKeyPair>` once, then performs
-   * O(1) membership lookups for each filter element. Faster than nested `any { contains(...) }` for
-   * batches with many entity keys.
-   *
-   * Caller must ensure `batch.entityKeys` is non-empty before invoking.
-   */
-  private fun batchEntityKeysOverlap(
-    batchEntityKeys: List<org.wfanet.measurement.edpaggregator.v1alpha.EntityKeyGroup>,
-    filter: Set<LabeledImpression.EntityKey>,
-  ): Boolean {
-    val batchKeyPairs: Set<EntityKeyPair> =
-      batchEntityKeys
-        .flatMap { g -> g.entityIdsList.map { id -> EntityKeyPair(g.entityType, id) } }
-        .toSet()
-    return filter.any { fk -> EntityKeyPair(fk.entityType, fk.entityId) in batchKeyPairs }
-  }
-
-  /** Flattened (entity_type, entity_id) pair for O(1) batch entity-key lookups. */
-  private data class EntityKeyPair(val entityType: String, val entityId: String)
 }
